refactor: split authentication and server store
