+++ packages/api-reference/src/components/Content/BaseUrl/BaseUrl.vue
@@ -3,7 +3,7 @@ import { findVariables } from '@scalar/api-client'
 import { ScalarIcon } from '@scalar/components'
 import { ref, watch } from 'vue'
 
-import { useGlobalStore } from '../../../stores'
+import { useServerStore } from '../../../stores'
 import { type Server, type Variable } from '../../../types'
 import { Card, CardContent, CardHeader } from '../../Card'
 import { MarkdownRenderer } from '../../MarkdownRenderer'
@@ -14,7 +14,7 @@ const props = defineProps<{
   value: Server[]
 }>()
 
-const { server, setServer } = useGlobalStore()
+const { server, setServer } = useServerStore()
 const selectedServerIndex = ref<number>(0)
 
 watch(
